requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.4.0
babel==2.17.0
beautifulsoup4==4.14.2
bleach==6.2.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
clip-anytorch==2.6.0
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
einops==0.8.1
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.20.0
fonttools==4.60.1
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2025.9.0
ftfy==6.3.1
gitdb==4.0.12
gitpython==3.1.45
grpcio==1.76.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
imageio==2.37.0
ipykernel==7.0.1
ipython==9.6.0
ipython-pygments-lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
jedi==0.19.2
jinja2==3.1.6
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-client==8.6.3
jupyter-console==6.6.3
jupyter-core==5.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter-server==2.17.0
jupyter-server-terminals==0.5.3
jupyterlab==4.4.10
jupyterlab-pygments==0.3.0
jupyterlab-server==2.28.0
jupyterlab-widgets==3.0.15
kiwisolver==1.4.9
lark==1.3.0
lazy-loader==0.4
lightning-utilities==0.15.2
lpips==0.1.4
markdown==3.9
markupsafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.2.1
mistune==3.1.4
mpmath==1.3.0
multidict==6.7.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
notebook==7.4.7
notebook-shim==0.2.4
numpy==2.3.4
opencv-python==4.11.0.86
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
pillow==12.0.0
platformdirs==4.5.0
prometheus-client==0.23.1
prompt-toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.0
psutil==7.1.1
pure-eval==0.2.3
pycparser==2.23
pydantic==2.12.3
pydantic-core==2.41.4
pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-json-logger==4.0.0
pytorch-lightning==2.5.5
pytz==2025.2
pywinpty==3.0.2
pyyaml==6.0.3
pyzmq==27.1.0
referencing==0.37.0
regex==2025.10.23
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.28.0
safetensors==0.6.2
scikit-image==0.25.2
scipy==1.16.2
seaborn==0.13.2
send2trash==1.8.3
sentencepiece==0.2.1
sentry-sdk==2.42.1
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
stack-data==0.6.3
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
terminado==0.18.1
tifffile==2025.10.16
tinycss2==1.4.0
tokenizers==0.22.1
torch==2.9.0
torchaudio==2.9.0
torchmetrics==1.8.2
torchvision==0.24.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.1
trimesh==4.9.0
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
uri-template==1.3.0
urllib3==2.5.0
uv-add-requirements-txt==1.5.2
wandb==0.22.2
wcwidth==0.2.14
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.9.0
werkzeug==3.1.3
widgetsnbextension==4.0.14
yarl==1.22.0
